@@ -2,8 +2,7 @@ package com.yehor.movieland.dao;
 
 import com.yehor.movieland.entity.Movie;
 
-import java.util.List;
-
 public interface MovieDao {
-    List<Movie> getAllMovies();
+    Iterable<Movie> findAll();
+    Iterable<Movie> getThreeRandomMovies();
 }
