@@ -2,22 +2,24 @@ package com.yehor.movieland.dao.jdbc.mapper;
 
 import com.yehor.movieland.entity.Movie;
 import org.springframework.jdbc.core.RowMapper;
+import org.springframework.stereotype.Component;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
+@Component
 public class MovieRowMapper implements RowMapper<Movie> {
 
     @Override
     public Movie mapRow(ResultSet resultSet, int i) throws SQLException {
-        Movie movie = new Movie();
-        movie.setId(resultSet.getInt("id"));
-        movie.setName_russian(resultSet.getString("name_russian"));
-        movie.setName_native(resultSet.getString("name_native"));
-        movie.setYear_of_release(resultSet.getDate("year_of_release"));
-        movie.setRating(resultSet.getInt("rating"));
-        movie.setPrice(resultSet.getInt("price"));
-        movie.setPicture_path(resultSet.getString("picture_path"));
-        return movie;
+        return Movie.builder()
+            .id(resultSet.getInt("id"))
+            .nameRussian(resultSet.getString("name_russian"))
+            .nameNative(resultSet.getString("name_native"))
+            .yearOfRelease(resultSet.getDate("year_of_release").toLocalDate())
+            .rating(resultSet.getDouble("rating"))
+            .price(resultSet.getDouble("price"))
+            .picturePath(resultSet.getString("picture_path"))
+            .build();
     }
 }
