@@ -2,8 +2,7 @@ package com.yehor.movieland.service;
 
 import com.yehor.movieland.entity.Movie;
 
-import java.util.List;
-
 public interface MovieService {
-    List<Movie> getAllMoviesJson();
+    Iterable<Movie> findAll();
+    Iterable<Movie> getThreeRandomMoviesJson();
 }
